requirements.txt
beautifulsoup4==4.10.0
discord.py==2.6.3
psutil==5.9.0
pygsheets==2.0.6
Requests==2.32.5
selenium==4.35.0
